requirements.txt
Flask
tensorflow
numpy
Pillow
matplotlib
roboflow
supervision
opencv-python
requests
inference-sdk